+++ TheMovieDb/View Controllers/Reviews/MVVM/Model/Models/Review.swift
@@ -1,0 +1,14 @@
+//
+//  Review.swift
+//  TheMovieDb
+//
+//  Created by developer on 30/11/21.
+//
+
+import Foundation
+
+struct Review: ReviewProtocol, Decodable {
+    var id: String?
+    var author: String
+    var content: String
+}
