+++ TheMovieDb/View Controllers/Reviews/MVVM/View Model/ReviewsViewModelList.swift
@@ -1,0 +1,13 @@
+//
+//  ReviewsViewModelList.swift
+//  TheMovieDb
+//
+//  Created by developer on 01/12/21.
+//
+
+import Foundation
+
+struct ReviewsViewModelList: ReviwsViewModelListProtocol {
+    typealias ReviewsTypes = ReviewViewModel
+    var results: [ReviewViewModel]
+}
