+++ TheMovieDb/View Controllers/Reviews/MVVM/View/Review Cell/ReviewCollectionViewCell.xib
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="19162" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="19144"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="ReviewCollectionViewCell" id="gTV-IL-0wX" customClass="ReviewCollectionViewCell" customModule="TheMovieDb" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="399" height="63"/>
+            <autoresizingMask key="autoresizingMask"/>
+            <view key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center">
+                <rect key="frame" x="0.0" y="0.0" width="399" height="63"/>
+                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                <subviews>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="xFX-6j-gTm">
+                        <rect key="frame" x="20" y="10" width="368" height="43"/>
+                        <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                </subviews>
+            </view>
+            <viewLayoutGuide key="safeArea" id="ZTg-uK-7eu"/>
+            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+            <constraints>
+                <constraint firstItem="xFX-6j-gTm" firstAttribute="centerY" secondItem="gTV-IL-0wX" secondAttribute="centerY" id="KpH-8W-mQI"/>
+                <constraint firstItem="xFX-6j-gTm" firstAttribute="leading" secondItem="gTV-IL-0wX" secondAttribute="leadingMargin" id="Uhy-uk-t7X"/>
+                <constraint firstAttribute="trailing" secondItem="xFX-6j-gTm" secondAttribute="trailing" constant="11" id="aEr-zi-ZSK"/>
+                <constraint firstItem="xFX-6j-gTm" firstAttribute="top" secondItem="gTV-IL-0wX" secondAttribute="top" constant="10" id="lwM-iP-0JQ"/>
+                <constraint firstAttribute="bottom" secondItem="xFX-6j-gTm" secondAttribute="bottom" constant="10" id="waS-4u-SqF"/>
+            </constraints>
+            <size key="customSize" width="399" height="63"/>
+            <connections>
+                <outlet property="reviewLabel" destination="xFX-6j-gTm" id="gVG-dB-19d"/>
+            </connections>
+            <point key="canvasLocation" x="390.57971014492756" y="94.754464285714278"/>
+        </collectionViewCell>
+    </objects>
+</document>
